+++ R/wdpa_fetch.R
@@ -69,13 +69,33 @@ NULL
 #' (UNEP-WCMC 2019).
 #'
 #' @section Troubleshooting:
-#' This function will sometimes return the error message
-#' `PhantomJS signals port = 4567 is already in use`.
-#' This error message can occur when you have previously run the function and
-#' it threw an error, or it terminated early.
-#' It can also occur when attempting to run the the function in multiple
-#' sessions on the same computer.
-#' To address this issue, you will need to restart your computer.
+#' The function requires a Chromium-based browser
+#' (e.g., Google Chrome, Chromium, or Brave) to be installed.
+#' This is because it uses the \pkg{chromote} to find the URL
+#' for downloading data from Protected Planet.
+#' If you don't have one of these browsers installed, then please try
+#' installing Google Chrome.
+#' If you do have one of these browsers installed and this function
+#' throws an error indicating that it can't find the browser,
+#' try setting the `CHROMOTE_CHROME` environment variable to the
+#' file path of the executable. For example, you could do this with:
+#' ````
+#' Sys.setenv(CHROMOTE_CHROME = "INSERT_FILE_PATH_HERE.exe")
+#' ````
+#'
+#' Also, the function will sometimes produce a message
+#' that complains about a `handle_read_frame` error. Please understand
+#' that this message is, in fact, not an error and can be safely ignored
+#' (see <https://github.com/rstudio/chromote/pull/111>).
+#' As such, if you see this message when running the function,
+#' you can assume that the function still worked correctly.
+#' For reference, the misleading message will look something like this:
+#' ````
+#' [2024-04-23 12:06:36] [error] handle_read_frame error: websocketpp.transport:7 (End of File)
+#' ```
+#'
+#' For further help with troubleshooting, please refer to the documentation
+#' for the \pkg{chromote} package (https://rstudio.github.io/chromote/).
 #'
 #' @return [sf::sf()] object.
 #'