+++ tests/testthat/helper_skip_functions.R
@@ -5,8 +5,8 @@ skip_if_local_and_slow_internet <- function(x) {
   testthat::skip_if(x, paste("On local and slow internet"))
 }
 
-skip_if_phantomjs_not_available <- function() {
+skip_if_chrome_not_available <- function() {
   testthat::skip_if_not(
-    is_phantomjs_available, "phantomJS not available"
+    is_chrome_installed, "chrome not available"
   )
 }