+++ tests/testthat/helper_variables.R
@@ -1,5 +1,5 @@
-# set default availability for phantomJS
-is_phantomjs_available <- FALSE
+# set default availability for chrome
+is_chrome_installed <- FALSE
 if (interactive() && identical(Sys.getenv("NOT_CRAN"), "true")) {
-  is_phantomjs_available <- TRUE
+  is_chrome_installed <- TRUE
 }