+++ tests/testthat/test_000_setup.R
@@ -1,14 +1,14 @@
 context("setup")
 
 # initialize data
-test_that("phantomjs", {
+test_that("initialize_chrome", {
   skip_on_cran()
-  # check if phantomjs installed and install if needed
-  if (!isTRUE(wdpar:::has_phantomjs())) {
+  # check if chrome installed and install if needed
+  if (!isTRUE(wdpar:::is_chrome_available())) {
     webdriver::install_phantomjs()
   }
   # update variable
-  is_phantomjs_available <<- wdpar:::has_phantomjs()
+  is_chrome_installed <<- wdpar:::is_chrome_available()
   # tests
-  expect_is(is_phantomjs_available, "logical")
+  expect_is(is_chrome_installed, "logical")
 })