+++ tests/testthat/test_wdpa_clean.R
@@ -15,7 +15,7 @@ default_retain_status <- c("Designated", "Inscribed", "Established")
 test_that("single country with eez", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -40,7 +40,7 @@ test_that("single country with eez", {
 test_that("single country without eez", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -62,7 +62,7 @@ test_that("single country without eez", {
 test_that("single country with simplification", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -84,7 +84,7 @@ test_that("single country with simplification", {
 test_that("single country without overlap removal", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -106,7 +106,7 @@ test_that("single country without overlap removal", {
 test_that("country with MULTIPOINT protected areas", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -122,7 +122,7 @@ test_that("country with MULTIPOINT protected areas", {
 test_that("country with MULTIPOLYGON protected area", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -140,7 +140,7 @@ test_that("country with MULTIPOLYGON protected area", {
 test_that("country with super invalid MULTIPOLYGON data", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- wdpa_fetch("GAB", wait = TRUE, check_version = FALSE)
@@ -152,7 +152,7 @@ test_that("country with super invalid MULTIPOLYGON data", {
 test_that("geometries in non-geometry column", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- wdpa_fetch("GAB", wait = TRUE, check_version = FALSE)
@@ -168,7 +168,7 @@ test_that("geometries in non-geometry column", {
 test_that("single country with no valid non-empty geometries", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- wdpa_clean(
@@ -192,7 +192,7 @@ test_that("single country with no valid non-empty geometries", {
 test_that("retain UNESCO Biosphere reserves", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -215,7 +215,7 @@ test_that("retain UNESCO Biosphere reserves", {
 test_that("protected areas that turn into long rectangles without prepr", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_if_not_installed("prepr")
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
@@ -231,7 +231,7 @@ test_that("protected areas that turn into long rectangles without prepr", {
 test_that("protected areas that massively increase in size without prepr", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_if_not_installed("prepr")
   skip_if_not_installed("dplyr")
   skip_on_github_workflow("Windows")
@@ -240,6 +240,8 @@ test_that("protected areas that massively increase in size without prepr", {
   ids <- c(23177, 12352, 555705343, 555705341, 555721495)
   x <- wdpa_fetch("DZA", wait = TRUE, check_version = FALSE)
   x <- x[x$WDPAID %in% ids, , drop = FALSE]
+  # skip none of the ids are present
+  if (nrow(x) < 1) skip(message = "data not available")
   # clean data
   y <- wdpa_clean(x, erase_overlaps = FALSE)
   # run tests
@@ -249,7 +251,7 @@ test_that("protected areas that massively increase in size without prepr", {
 test_that("custom retain_status", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -272,7 +274,7 @@ test_that("custom retain_status", {
 test_that("NULL retain_status", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data
@@ -293,7 +295,7 @@ test_that("NULL retain_status", {
 test_that("empty intersections", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data