+++ tests/testthat/test_wdpa_fetch.R
@@ -3,7 +3,7 @@ context("wdpa_fetch")
 test_that("country name", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- suppressWarnings(wdpa_fetch("Liechtenstein", wait = TRUE))
@@ -14,7 +14,7 @@ test_that("country name", {
 test_that("ISO3", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- suppressWarnings(wdpa_fetch("LIE", wait = TRUE, verbose = TRUE))
@@ -26,7 +26,7 @@ test_that("global", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
   skip_if_local_and_slow_internet()
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   x <- suppressWarnings(wdpa_fetch(
@@ -37,7 +37,7 @@ test_that("global", {
 test_that("polygon and point data", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_if_local_and_slow_internet()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
@@ -54,7 +54,7 @@ test_that("polygon and point data", {
 test_that("cache", {
   skip_on_cran()
   skip_if_not(curl::has_internet())
-  skip_if_phantomjs_not_available()
+  skip_if_chrome_not_available()
   skip_on_github_workflow("Windows")
   skip_on_github_workflow("macOS")
   # fetch data